Addressing issue #10. Cleaned up code, moved mock test classes into their own files.
