+++ tests/mock/abstractions/AbstractImplementationOfInterfaceForClassThatDefinesMethods.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Darling\PHPMockingUtilities\tests\mock\abstractions;
+
+use \Darling\PHPMockingUtilities\tests\mock\interfaces\InterfaceForClassThatDefinesMethods;
+
+/**
+ * The following classes are defined here for use by
+ * the MockClassInstanceTest
+ *
+ */
+
+abstract class AbstractImplementationOfInterfaceForClassThatDefinesMethods
+{
+}
