+++ tests/mock/classes/ClassThatDoesNotDefineMethods.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Darling\PHPMockingUtilities\tests\mock\classes;
+
+class ClassThatDoesNotDefineMethods
+{
+
+}
